consolidate collaboration assets as suggested
