+++ src/modules/toolbox/components/drawToolContent/DrawToolContent.js
@@ -16,7 +16,7 @@ import { FileStorageState } from '../../../../store/fileStorage/fileStorage.redu
 import loadingSvg from './assets/cloud-arrow.svg';
 import cloudCheckSvg from './assets/cloud-check.svg';
 import recordCircleSvg from './assets/cloud-slash.svg';
-import peopleSvg from './assets/people.svg';
+import peopleSvg from '../../../../assets/icons/people.svg';
 
 const Update = 'update';
 const Update_Tools = 'update_tools';
